Make modifications to allow for easier postgres conversion
- Surround everything with `"` marks, not `\`` marks.
- Alias all subqueries
- Aliases surrounded with " marks, not `[` and `]`

This is not something that I would ever consider merging back into the
origin, as these are not things that are generic for all SQLite uses,
but just for our needs. There are certain times when we are unable to
parse a SQLite query using the Postgres parser due to differences in
syntax (such as SQLite not requiring a subquery to be aliased).